TheSoapMaker/soapmaker-backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
dj-rest-auth[with_social]>=6.0
django-allauth>=0.62
drf-spectacular>=0.27
django-filter>=24.1
cloudinary>=1.40
django-cloudinary-storage>=0.3
stripe>=8.0
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2
